small CC tests refactor
